fix: 경험치 로직 수정

+++ bookduck/src/main/java/com/mmc/bookduck/domain/user/dto/response/UserGrowthInfoResponseDto.java
@@ -1,7 +1,7 @@
 package com.mmc.bookduck.domain.user.dto.response;
 
 public record UserGrowthInfoResponseDto (
-        int level,
+        long level,
         long expInCurrentLevel,
         long expToNextLevel
 ) {
